deployment/requirements.txt
python-jenkins==1.7.0